src/server/mcp_hub/gtasks/requirements.txt
# Create new file: src/server/mcp_hub/gtasks/requirements.txt
fastmcp
python-dotenv
motor
cryptography
composio